refactor(Lotto): 패키지 이동
